requirements.txt
pandas
numpy
scipy
flask
python-dotenv
groq